Update consumer components to use usePostMeta
