+++ projects/js-packages/publicize-components/src/components/share-buttons/useShareButtonText.ts
@@ -1,6 +1,6 @@
 import { useSelect } from '@wordpress/data';
 import { useCallback } from '@wordpress/element';
-import { getShareMessage } from '../../utils';
+import { usePostMeta } from '../../utils';
 
 /**
  * Prepares the text to share.
@@ -8,19 +8,23 @@ import { getShareMessage } from '../../utils';
  * @returns {(textWithPlaceholders: string, isUrl: boolean) => string} A function that accepts the text with placeholders and returns the text with the placeholders replaced.
  */
 export function useShareButtonText() {
-	const { message, link } = useSelect( select => {
-		// eslint-disable-next-line @typescript-eslint/no-explicit-any
-		const getEditedPostAttribute = ( select( 'core/editor' ) as any )
-			.getEditedPostAttribute satisfies ( attribute: string ) => unknown;
+	const { shareMessage } = usePostMeta();
+	const { message, link } = useSelect(
+		select => {
+			// eslint-disable-next-line @typescript-eslint/no-explicit-any
+			const getEditedPostAttribute = ( select( 'core/editor' ) as any )
+				.getEditedPostAttribute satisfies ( attribute: string ) => unknown;
 
-		return {
-			link: getEditedPostAttribute( 'link' ),
-			message:
-				getShareMessage() ||
-				getEditedPostAttribute( 'meta' )?.jetpack_seo_html_title ||
-				getEditedPostAttribute( 'title' ),
-		};
-	}, [] );
+			return {
+				link: getEditedPostAttribute( 'link' ),
+				message:
+					shareMessage ||
+					getEditedPostAttribute( 'meta' )?.jetpack_seo_html_title ||
+					getEditedPostAttribute( 'title' ),
+			};
+		},
+		[ shareMessage ]
+	);
 
 	return useCallback(
 		( textWithPlaceholders: string, isUrl = true ) => {
